TFB-010 - some refactoring was made (OAuth2Service added, field 'email' deleted from AuthProvider entity etc.)
