+++ backend/src/main/java/com/bookamore/backend/entity/enums/ProviderType.java
@@ -1,5 +1,5 @@
 package com.bookamore.backend.entity.enums;
 
 public enum ProviderType {
-    GOOGLE, FACEBOOK, DISCORD, GITHUB
+    GOOGLE, FACEBOOK
 }
