+++ backend/src/main/java/com/bookamore/backend/service/OAuth2Service.java
@@ -1,0 +1,11 @@
+package com.bookamore.backend.service;
+
+import com.bookamore.backend.entity.User;
+import com.bookamore.backend.entity.enums.ProviderType;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+public interface OAuth2Service {
+    User findOrCreateOAuth2User(ProviderType provider,
+                                String providerUserId,
+                                OAuth2User oauth2User);
+}
