+++ backend/src/main/resources/.env.origin
@@ -1,3 +1,7 @@
 JWT_SECRET=
 JWT_EXPIRATION=
 CLIENT_URL=
+GOOGLE_CLIENT_ID=
+GOOGLE_CLIENT_SECRET=
+FACEBOOK_CLIENT_ID=
+FACEBOOK_CLIENT_SECRET=
